+++ wayland-backend/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### Additions
+- client: Implement `Eq` for `Backend`
+
 ## 0.3.2 -- 2023-09-25
 
 #### Bugfixes