+++ wayland-client/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+#### Additions
+- Implement `Eq` for `Connection`
+
 ## 0.31.1 -- 2023-09-19
 
 #### Additions